Create curl_malware_downloader.sh
Tool designed in bash to download the contents of a specified web page using multiple different predefined user agent strings, or a singular user agent string, which will poll the site every X seconds/minutes.

The time delay adds in a small randomisation factor between 0-30% of the expected time to make it appear pseudo-random to the recipient host.

Once the downloads have occurred, they are saved into a local folder structure and compared using various hashing tools. A landing page is generated which will automatically be opened upon completion.

The header and body content are split up and saved as .txt files in an attempt to stop the browser from rendering them as they can be viewed directly from the landing page.

Bash is a dirty language. I am not a programmer by trade. My code is as neat as *I* can make it, so any hints or tips to clean up the code are welcome.

This program has received minimal testing, but we have decided to upload it because...well just because :)