Create cirt_file_analysis.sh
A tool designed for rapid static file analysis. 

There are more in depth tools, but this one has no dependencies on a standard ubuntu 14.04 install and also outputs into a html structure which makes it nice and easy to read.

Its far from perfect, and as always, I am not a programmer by trade. Please feel free to advise me on any tips or techniques that can help me better my code.