Create plugin_explanation.txt
This goes with volnalysis so that the outputs pages can display informational content. This was created after my SANS 526 and used to help me better understand what was occurring on the back end.

I wrote the content in my own time but some of it is pretty rough, so I'll leave it to you to fill in your own, although I have filled an example out from my notes.

Its far from perfect, and as always, I am not a programmer by trade. Please feel free to advise me on any tips or techniques that can help me better my code.